@@ -5,14 +5,20 @@ USAGE:
     $ idpconfgen fastaext [PDBS]
 """
 import argparse
-import itertools as it
+from functools import partial
 
 from idpconfgen import Path, log
-from idpconfgen.libs import libcli, libio, libpdb, libstructure
-from idpconfgen.logger import S, T, init_files
+from idpconfgen.libs import libcli
+from idpconfgen.libs.libio import FileReaderIterator, save_dictionary
+from idpconfgen.libs.libmulticore import pool_function
+from idpconfgen.libs.libpdb import get_fasta_from_PDB
+from idpconfgen.logger import T, init_files
 
 
 LOGFILESNAME = 'idpconfgen_fastaext'
+
+_name = 'fastaext'
+_help = 'Extract FASTA sequence from PDBs.'
 
 _prog, _des, _us = libcli.parse_doc_params(__doc__)
 
@@ -22,13 +28,8 @@ ap = libcli.CustomParser(
     usage=_us,
     formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-# https://stackoverflow.com/questions/24180527
 
-ap.add_argument(
-    'pdbs',
-    help='PDB file list.',
-    nargs='+',
-    )
+libcli.add_argument_pdb_files(ap)
 
 ap.add_argument(
     '-o',
@@ -40,27 +41,17 @@ ap.add_argument(
         ),
     type=Path,
     default=None,
+    const='fasta.json',
+    action=libcli.CheckExt({'.json'}),
     )
 
-
-def _load_args():
-    cmd = ap.parse_args()
-    return cmd
-
-
-def maincli():
-    """
-    Execute main client function.
-
-    Reads command line arguments and executes logic.
-    """
-    cmd = _load_args()
-    main(**vars(cmd))
+libcli.add_argument_ncores(ap)
 
 
 def main(
-        pdbs,
+        pdb_files,
         output=None,
+        ncores=1,
         **kwargs,
         ):
     """
@@ -79,23 +70,20 @@ def main(
     init_files(log, LOGFILESNAME)
 
     log.info(T('reading input paths'))
-    # tee is used to keep memory footprint low
-    # though it would be faster to create a list from path_bundle
-    _, pdbs = it.tee(libio.read_path_bundle(pdbs))
-    log.info(S('done'))
-    pdbids = libpdb.PDBList(_)
+    pdbs2operate = FileReaderIterator(pdb_files, ext='.pdb')
 
-    out_data = []
-    for pdbid, pdbfile in zip(pdbids, pdbs):
-        structure = libstructure.Structure(pdbfile)
-        structure.build()
-        fasta = structure.fasta
-        assert len(fasta) == 1
-        out_data.append('{}|{}'.format(pdbid, next(iter(fasta.values()))))
+    execute = partial(
+        pool_function,
+        get_fasta_from_PDB,
+        pdbs2operate,
+        ncores=ncores,
+        )
 
-    libio.write_text('\n'.join(out_data), output)
+    mdict = {i: j for i, j in execute()}
+
+    save_dictionary(mdict, output)
     return
 
 
 if __name__ == '__main__':
-    maincli()
+    libcli.maincli(ap, main)
