@@ -1,20 +1,35 @@
 """
-Extracts secondary structure information from PDBs.
+Extract secondary structure protein segments from PDBs.
 
-Uses an external third party software.
+Requires a *.dssp file as generated by `idpconfgen ssext` CLI.
 
 USAGE:
-    $ idpconfgen ssext [PDBS]
+
+idpcfongen segext PDBS_FOLDER DSSPFILE -d OUTPUTFOLDER -s [L/H/E/A]
 """
 import argparse
+from functools import partial
 
-from idpconfgen import Path, log
-from idpconfgen.libs import libcli, libio, libmulticore, libparse
+from idpconfgen import log
+from idpconfgen.libs import libcli
 from idpconfgen.logger import S, T, init_files
+from idpconfgen.libs.libhigherlevel import extract_secondary_structure
+from idpconfgen.libs.libio import (
+    FileReaderIterator,
+    read_dictionary_from_disk,
+    save_pairs_to_disk,
+    )
+from idpconfgen.libs.libmulticore import (
+    consume_iterable_in_list,
+    flat_results_from_chunk,
+    pool_function_in_chunks,
+    )
 
 
-LOGFILESNAME = 'idpconfgen_ssext'
+LOGFILESNAME = '.idpconfgen_ssext'
 
+_name = 'ssext'
+_help = 'Extract secondary structure elements from PDBs.'
 _prog, _des, _us = libcli.parse_doc_params(__doc__)
 
 ap = libcli.CustomParser(
@@ -23,118 +38,120 @@ ap = libcli.CustomParser(
     usage=_us,
     formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-# https://stackoverflow.com/questions/24180527
+
+libcli.add_argument_pdb_files(ap)
 
 ap.add_argument(
-    'ss_cmd',
-    help='The path to the DSSP executable file.',
-    type=str,
+    'sscalc_file',
+    help='The DSSP file as saved by IDPConfGen SSCALC CLI',
     )
 
+libcli.add_argument_destination_folder(ap)
+
 ap.add_argument(
-    'pdbs',
-    help='PDB file list.',
+    '-s',
+    '--structure',
+    help=(
+        'The secondary structure type to extract. '
+        'Defaults to \'all\'. '
+        'A subfolder is created for each secondary structure type'
+        ),
+    default='all',
     nargs='+',
     )
 
 ap.add_argument(
-    '-o',
-    '--output',
+    '-a',
+    '--atoms',
     help=(
-        "The output file containing the PDBID and "
-        "respective secondary structure information. "
-        "Defaults to sys.stdout (prints to console)."
+        'List of atom names to save in the selection.\n'
+        'Defaults to `N`, `CA`, and `C`.'
         ),
-    type=Path,
-    default=None,
+    default='all',  # ('N', 'CA', 'C'),
+    nargs='+',
     )
-
-ap.add_argument(
-    '-r',
-    '--reduced',
-    help=(
-        'Reduces nomenclature for secondary structure identity '
-        'to \'L\', \'H\' and \'E\'.'
-        ),
-    action='store_true',
-    )
-
-ap.add_argument(
-    '-n',
-    '--ncores',
-    help='Number of cores to use',
-    default=1,
-    type=int,
-    )
-
-
-def _load_args():
-    cmd = ap.parse_args()
-    return cmd
-
-
-def maincli():
-    """
-    Execute main client function.
-
-    Reads command line arguments and executes logic.
-    """
-    cmd = _load_args()
-    main(**vars(cmd))
+libcli.add_argument_minimum(ap)
+libcli.add_argument_ncores(ap)
 
 
 def main(
-        ss_cmd,
-        pdbs,
-        output=None,
+        pdb_files,
+        sscalc_file,
+        atoms='all',
+        chunks=5000,
+        destination=None,
+        func=None,
+        minimum=4,
         ncores=1,
-        reduced=False,
-        **kwargs,
+        structure='all',
         ):
     """
-    Run main cli logic.
+    Extract secondary structure segments from PDBs.
 
     Parameters
     ----------
-    ss_cmd : str or Path
-        The command to run with subprocess module.
+    pdb_files : str or list
+        Paths to PDB files.
 
-    pdbs : list
-        A list of paths to PDB files or PDB file lists.
+    sscalc_file : str or Path
+        Path to the `sscalc` file containing PDBID codes and `dssp`
+        information.
 
-    output : string or Path, optional
-        If given prints output to that file, else prints to console.
-        Defaults to `None`.
+    atoms : str or list
+        Atoms to keep in the extracted segments.
+
+    minimum : int
+        The minimum residue length of the segments.
+        Segments with less than minimum are discarded.
+
+    destination : str or Path
+        Where to save the new PDBs.
+
+    structure : str or list
+        Secondary structure characters to consider.
+
+    chunks : int
+        The size of the chunk to process in memory before saving to
+        the disk.
 
     ncores : int
-        The numbers of cores to use.
+        The number of cores to use.
+        Defaults to 1.
     """
-    log.info(T('Extracting Secondary structure information'))
+    log.info(T('Extracting secondary structure elements.'))
     init_files(log, LOGFILESNAME)
 
-    log.info(T('reading input paths'))
-    pdbs = libio.read_path_bundle(pdbs)
-    log.info(S('done'))
+    ssdata = read_dictionary_from_disk(sscalc_file)
+    log.info(S('read sscalc file'))
+    pdbs2operate = FileReaderIterator(pdb_files, ext='.pdb')
+    log.info(S('read PDB files'))
 
-    log.info(T('preparing task execution'))
-    log.info(S('for {} cores', ncores))
-    ss_ext_exec = libmulticore.JoinedResults(
-        pdbs,
-        ss_cmd,
+    execute = partial(
+        # multiprocessing function
+        pool_function_in_chunks,
+        # wrapper around generator
+        consume_iterable_in_list,
+        # items
+        pdbs2operate,
+        # func to operate
+        extract_secondary_structure,
+        # multicore kwargs
         ncores=ncores,
-        TaskMethod=libmulticore.DSSPTask,
-        results_parser=libparse.DSSPParser.from_data_id_tuple,
-        reduced=reduced,
+        chunks=chunks,
+        # func kwargs
+        atoms=atoms,
+        ssdata=ssdata,
+        structure=structure,
+        minimum=minimum,
         )
-    log.info(S('executing...'))
-    ss_ext_exec.run()
 
-    log.info(T('exporting output to: {}', output))
-    libparse.export_ss_from_DSSP(*ss_ext_exec.results, output=output)
-
-    log.info(S('All done. Thanks!'))
+    flat_results_from_chunk(
+        execute,
+        save_pairs_to_disk,
+        destination=destination,
+        )
     return
 
 
 if __name__ == '__main__':
-    maincli()
+    libcli.maincli()
