@@ -18,12 +18,20 @@ class CIFParser:
     """
 
     __slots__ = [
+        '_auth_label',
         '_len',
         '_line',
         'cif_dict',
         ]
 
-    def __init__(self, datastr):
+    def __init__(self, datastr, label_priority='auth'):
+
+        _label_priority = {
+            'auth': self._auth_label_priority,
+            'label': self._label_auth_priority,
+            }
+
+        self._auth_label = _label_priority[label_priority]
         self.cif_dict = {}
         self.read_cif(datastr)
         self.line = 0
@@ -95,7 +103,13 @@ class CIFParser:
         """Retrieve the label value for the current :attr:`line`."""
         return self.cif_dict[label][self.line]
 
-    def _auth_label(self, label):
+    def _auth_label_priority(self, label):
+        try:
+            return self.get_value(f'_atom_site.auth_{label}')
+        except KeyError:
+            return self.get_value(f'_atom_site.label_{label}')
+
+    def _label_auth_priority(self, label):
         try:
             return self.get_value(f'_atom_site.label_{label}')
         except KeyError:
@@ -138,7 +152,7 @@ class CIFParser:
         """The chainID field of the current :attr:`line`."""  # noqa: D401
         value = self._auth_label('asym_id')
         if value in ('?', '.'):
-            value = self.get_value(f'_atom_site.auth_asym_id')
+            value = self.get_value('_atom_site.auth_asym_id')
         return value
 
     @property
@@ -279,9 +293,8 @@ def populate_cif_dictionary(lines, start_index, cif_dict):
     """
     assert len(lines) > start_index
 
-    valid_len = len(cif_dict.keys())
-    counter = 0
-    for line in lines[start_index:]:
+    valid_len = len(cif_dict)
+    for counter, line in enumerate(lines[start_index:]):
         if line.startswith('#'):
             return counter
 
@@ -298,8 +311,6 @@ def populate_cif_dictionary(lines, start_index, cif_dict):
 
         for i, key in enumerate(cif_dict.keys()):
             cif_dict[key].append(ls[i])
-
-        counter += 1
 
 
 def parse_cif_line(line):
