@@ -5,7 +5,7 @@ import logging
 class TitleLog:
     """
     Format string to title.
-    
+
     Example
     -------
         >>> TitleLog('error msg: {} {} {}', var1, var2 var3)
@@ -14,7 +14,7 @@ class TitleLog:
     def __init__(self, msg, *args):
         self.msg = msg.title()
         self.args = args
-    
+
     def __str__(self):
         """Represent object as string."""
         return '\n* {} ...'.format(self.msg.format(*self.args))
@@ -23,19 +23,19 @@ class TitleLog:
 class SubLog:
     """
     Format string to bullet point like structure.
-    
+
     This format performs nicely under the :class:`TitleLog` formatting.
-    
+
     Example
     -------
         >>> SubLog('error msg: {} {} {}', var1, var2 var3)
     """
-    
+
     def __init__(self, msg, *args, spacer=' ', indent=4):
         self.msg = msg
         self.args = args
         self.indent = spacer * indent
-    
+
     def __str__(self):
         """Represent object as string."""
         return '{}{}'.format(
@@ -51,7 +51,7 @@ S = SubLog
 def init_files(log, logfilesname):
     """
     Initiate log files.
-    
+
     Three log files are created:
         - .debug
         - .log
