@@ -1,34 +1,66 @@
 """Common funcs and variables for tests."""
 import random
+import shutil
+from contextlib import suppress
 
 from idpconfgen import Path
 
 
+class DummyFolder:
+    def __init__(self, folder):
+        self.folder = Path(folder)
+
+    def __enter__(self):
+        self.folder.mkdir(parents=True, exist_ok=True)
+        return self.folder
+
+    def __exit__(self, *args, **kwargs):
+        with suppress(FileNotFoundError):
+            shutil.rmtree(self.folder)
+
+
+class DummyFile:
+    def __init__(self, fpath):
+        self.fpath = Path(fpath)
+
+    def __enter__(self):
+        return self.fpath
+
+    def __exit__(self, *args, **kwargs):
+        with suppress(FileNotFoundError):
+            self.fpath.unlink()
+
+
+# folders
 tests_folder = Path(__file__).myparents()
-
-project_folder = tests_folder.parents[0]
+project_folder = tests_folder.parent
 iofiles_folder = Path(tests_folder, 'readfiles')
+folder_output = Path(tests_folder, 'data_out_scratch')
 data_folder = Path(tests_folder, 'data')
-folder_output = Path(
-    tests_folder,
-    'data_out_scratch',
-    )
+pdbids_folder = Path(data_folder, 'pdbids')
 
-
-cull = Path(data_folder, 'cull.list')
-pdb_example = Path(data_folder, 'pdb_example.pdb')
-pdb_models = Path(data_folder, 'pdb_models.pdb')
-pdb_models_output = Path(data_folder, 'pdb_models_output.pdb')
-PFA1 = Path(data_folder, 'PFA1_A.pdb')
-PFA2 = Path(data_folder, 'PFA2_A.pdb')
-PFA3 = Path(data_folder, 'PFA3_A.pdb')
-pdbs_fasta = Path(data_folder, 'pdbs.fasta')
+# files
+PFA1 = Path(pdbids_folder, 'PFA1_A.pdb')
+PFA2 = Path(pdbids_folder, 'PFA2_A.pdb')
+PFA3 = Path(pdbids_folder, 'PFA3_A.pdb')
+cif_EOF = Path(data_folder, 'cif_EOF.cif')
 cif_example = Path(data_folder, 'cif_example.cif')
 cif_example_auth = Path(data_folder, 'cif_example_auth.cif')
 cif_example_noasymid = Path(data_folder, 'cif_example_noasymid.cif')
 cif_noatomsite = Path(data_folder, 'cif_no_atom_site.cif')
 cif_nohash = Path(data_folder, 'cif_nohash.cif')
-cif_EOF = Path(data_folder, 'cif_EOF.cif')
+cull = Path(data_folder, 'cull.list')
+dict1json = Path(data_folder, 'dict1.json')
+example_dssp = Path(data_folder, 'example.dssp')
+file_tar = Path(data_folder, 'files.tar')
+pdb_example = Path(data_folder, 'pdb_example.pdb')
+pdb_models = Path(data_folder, 'pdb_models.pdb')
+pdb_models_output = Path(data_folder, 'pdb_models_output.pdb')
+pdb_res_gap = Path(data_folder, 'pdb_residue_gap.pdb')
+pdbs_fasta = Path(data_folder, 'pdbs.fasta')
+EXPL_A = Path(data_folder, 'EXPL_A.pdb')
+EXPL_B = Path(data_folder, 'EXPL_B.pdb')
+
 
 cif_example_headers = [
     '_atom_site.group_PDB',
