@@ -1,249 +1,221 @@
-"""Test data file parsers."""
+"""Test libparse module."""
+from pathlib import Path as Path_
+
 import pytest
 
 from idpconfgen import Path
-from idpconfgen.core import exceptions as EXCPTS
-from idpconfgen.libs import libparse, libpdb
+from idpconfgen.libs import libparse
 
 from . import tcommons
 
 
-class TestLITA():
-    """Test list index to array."""
+def test_t2s_dispacher():
+    """
+    Test type2string dispacher.
 
-    data = [
-        'i like to program in python',
-        'u like to program in java',
-        'w like to program in csharp',
+    All values should return a string.
+    """
+    types = {
+        str: 'hello',
+        bytes: b'world',
+        type(Path()): Path(__file__),
+        type(Path_()): Path(__file__),
+        }
+
+    for k, v in types.items():
+        assert isinstance(libparse.type2string[k](v), str)
+
+
+@pytest.mark.parametrize(
+    'in1, expected',
+    [
+        (
+            'LLLHEEE',
+            [['L', slice(0, 3)], ['H', slice(3, 4)], ['E', slice(4, 7)]]
+            ),
+        ]
+    )
+def test_group_by(in1, expected):
+    """Group sequence by equality."""
+    result = libparse.group_by(in1)
+    assert result == expected
+
+
+def test_group_runs_1():
+    """Test group runs."""
+    num = [-2, -1, 0, 1, 2, 3, 5, 6, 7]
+    expected = [
+        [-2, -1, 0, 1, 2, 3],
+        [5, 6, 7],
         ]
 
-    def test_list_index_to_array_1(self):
-        """Test index."""
-        parsed = libparse.list_index_to_array(self.data, index=0)
-        assert list(parsed) == ['i', 'u', 'w']
+    for i, result in enumerate(libparse.group_runs(num)):
+        assert result == expected[i]
+    else:
+        assert i == 1
 
-    def test_list_index_to_array_2(self):
-        """Test slice."""
-        parsed = libparse.list_index_to_array(self.data, start=21, stop=25)
-        assert list(parsed) == ['pyth', 'java', 'csha']
 
-    def test_list_index_to_array_3(self):
-        """Test ValueError."""
-        with pytest.raises(ValueError):
-            libparse.list_index_to_array(self.data)
-   
-    def test_list_index_to_array_4(self):
-        """Test IndexError."""
+def test_group_runs_2():
+    """Test group runs."""
+    num = [-2, -1, 0, 1, 2, 3, 5, 6, 7]
+    expected = [num]
+
+    for i, result in enumerate(libparse.group_runs(num, tolerance=2)):
+        assert result == expected[i]
+    else:
+        assert i == 0
+
+
+def test_parse_dssp():
+    """Parse DSSP from example."""
+    results = [
+        (
+            '  S      T ',
+            'SRMPSPPMPSS',
+            [str(i) for i in (list(range(47, 56)) + [348, 349])],
+            ),
+        (
+            '   EEEE ',
+            'AAPRLSFL',
+            [str(i) for i in range(395, 403)],
+            ),
+        (
+            '    ',
+            'XXXX',
+            [str(i) for i in range(501, 505)],
+            ),
+        ]
+
+    with tcommons.example_dssp.open('r') as dssp_data:
+        for i, result in enumerate(libparse.parse_dssp(dssp_data.read())):
+            assert result == results[i]
+        else:
+            assert i == 2
+
+
+def test_parse_dssp_indexError():
+    """
+    Test IndexError.
+
+    Obtained in case `dssp` file as not the `#` sperator.
+    """
+    with open(Path(tcommons.data_folder, 'wrong.dssp')) as fin:
+        data = fin.read()
         with pytest.raises(IndexError):
-            libparse.list_index_to_array(self.data, index=400)
-
-    def test_list_index_to_array5(self):
-        """Test slice object."""
-        parsed = libparse.list_index_to_array(self.data, sObj=slice(21, 25, 1))
-        assert list(parsed) == ['pyth', 'java', 'csha']
-    
-    def test_list_index_to_array6(self):
-        """Test slice object."""
-        with pytest.raises(ValueError):
-            libparse.list_index_to_array(self.data, sObj=5)
+            for _ in libparse.parse_dssp(data):
+                pass
 
 
-class TestDSSPParser:
-    """Test DSSPParser."""
-
-    dssp_file = tcommons.data_folder / '1ABC_D.dssp'
-    dssp_file2 = tcommons.data_folder / '1ABC_E.dssp'
-    dssp_wrong = tcommons.data_folder / 'wrong.dssp'
-    dssp_wrong2 = tcommons.data_folder / 'wrong2.dssp'
-    dsspdata = dssp_file.read_text()
-
-    def test_dssp1(self):
-        """Test initiation empty args."""
-        libparse.DSSPParser()
-
-    def test_dssp1_1(self):
-        """Test pdbid attribute."""
-        dobj = libparse.DSSPParser(pdbid='1ABC')
-        assert dobj.pdbid == '1ABC'
-
-    def test_dssp2(self):
-        """Test initiation from data."""
-        libparse.DSSPParser(data=self.dsspdata)
-
-    def test_dssp3(self):
-        """Test initiation from file."""
-        libparse.DSSPParser(fin=self.dssp_file)
-
-    def test_dssp3_1(self):
-        """Test initiation from data."""
-        lines = Path(self.dssp_file).read_text().split('\n')
-        libparse.DSSPParser(data=lines)
-
-    def test_dssp4(self):
-        """Test secondary structure parsing."""
-        dobj = libparse.DSSPParser(fin=self.dssp_file)
-
-        expected = [' ', 'H', 'G', 'I', 'B', 'E', ' ', 'T', 'S', ' ']
-        assert list(dobj.ss) == expected
+@pytest.mark.parametrize(
+    'in1, expected',
+    [
+        ('A', {'A', 'a'}),
+        ('aa', {'AA', 'Aa', 'aA', 'aa'}),
+        ('a_', {'A_', 'a_'}),
+        ]
+    )
+def test_case_combinations(in1, expected):
+    """Test cases combinations."""
+    result = libparse.sample_case(in1)
+    assert result == expected
 
 
-    def test_dssp4(self):
-        """Test secondary structure parsing."""
-        dobj = libparse.DSSPParser(fin=self.dssp_file, reduced=True)
-        expected = ['L', 'H', 'H', 'H', 'E', 'E', 'L', 'L', 'L', 'L']
-        assert list(dobj.ss) == expected
-
-
-    def test_dssp5(self):
-        """Test DSSPParserError."""
-        with pytest.raises(EXCPTS.DSSPParserError):
-            libparse.DSSPParser(fin=self.dssp_wrong)
-
-    def test_dssp6(self):
-        """Test find index."""
-        data = self.dssp_file.read_text().split('\n')
-        index = libparse.DSSPParser._finds_data_index(data)
-        assert index == 28
-
-    def test_dssp7(self):
-        """
-        Test wrong ss capturing.
-
-        The input here has the '#' wrongly placed.
-        Captures if any of the registered chars does not belong
-        to the possible DSSP secondary structure chars.
-        """
-        with pytest.raises(EXCPTS.DSSPSecStructError):
-            libparse.DSSPParser(fin=self.dssp_wrong2)
-
-    def test_dssp8(self):
-        """Test file does not exist exception."""
-        with pytest.raises(FileNotFoundError):
-            libparse.DSSPParser(fin='doesnotexist.dssp')
-
-    def test_dssp9(self):
-        """Test FASTA extraction."""
-        dobj = libparse.DSSPParser(fin=self.dssp_file)
-        expected = ['K', 'K', 'V', 'K', 'V', 'S', 'H', 'R', 'S', 'H']
-        assert list(dobj.fasta) == expected
-
-    def test_dssp10(self):
-        """Test equality with pdbid None."""
-        dobj1 = libparse.DSSPParser(fin=self.dssp_file)
-        dobj2 = libparse.DSSPParser(fin=self.dssp_file)
-
-        assert dobj1 == dobj2
-
-    def test_dssp11(self):
-        """Test equality with pdbid."""
-        dobj1 = libparse.DSSPParser(fin=self.dssp_file, pdbid='1ABC')
-        dobj2 = libparse.DSSPParser(fin=self.dssp_file, pdbid='1ABC')
-
-        assert dobj1 == dobj2
-    
-    def test_dssp12(self):
-        """Test differ with pdbid different."""
-        dobj1 = libparse.DSSPParser(fin=self.dssp_file, pdbid='1ABC')
-        dobj2 = libparse.DSSPParser(fin=self.dssp_file, pdbid='1XXX')
-
-        assert dobj1 != dobj2
-
-    def test_dssp13(self):
-        """Test differ with data different."""
-        dobj1 = libparse.DSSPParser(fin=self.dssp_file, pdbid='1ABC')
-        dobj2 = libparse.DSSPParser(fin=self.dssp_file2, pdbid='1ABC')
-
-        assert dobj1 != dobj2
-
-
-class TestDSSPMediator:
+@pytest.fixture
+def BTE_A_results():
     """
-    Test DSSP Mediator.
-    
-    DSSPMediator takes the output from the DSSPTask and passes it
-    to the DSSPParser.
+    Return the expected dictionary for split_pdb_by_dssp.
+
+    Considers using the `reduced` kwarg.
     """
-
-    dssp_file = Path(tcommons.data_folder, '1ABC_D.dssp')
-    pdb_file = Path(tcommons.data_folder, '1A12.pdb')
-    dsspdata = dssp_file.read_text()
-
-    def test_dsspmediator_0(self):
-        """Test DSSPParser is returned."""
-        dssp_mediated = libparse.DSSPParser.from_data_id_tuple((
-            self.pdb_file,
-            self.dsspdata,
-            ))
-        assert isinstance(dssp_mediated, libparse.DSSPParser)
-    
-    def test_dsspmediator_1(self):
-        """Test mediated is of correct type and form."""
-        dssp_parser = libparse.DSSPParser(
-            data=self.dsspdata,
-            pdbid=libpdb.PDBIDFactory(self.pdb_file),
-            )
-
-        dssp_mediated = libparse.DSSPParser.from_data_id_tuple((
-            self.pdb_file,
-            self.dsspdata,
-            ))
-
-        assert dssp_parser == dssp_mediated
+    results = {
+        "1BTE_A_seg0": {
+            "dssp": "LLLEEEEEELLHHHHLLLLEEEEELLL",
+            "fasta": "ETQECLFFNANWERDRTNQTGVEPCYG",
+            "resids": "7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33",  # noqa: E501
+            },
+        "1BTE_A_seg1": {
+            "dssp": "LLEEEEEEEEELLEEEEEEEEEELLLLLLLLLLLEEELLLLLLLEEEEELLLLHHHLEEELLLL",  # noqa: E501
+            "fasta": "RRHCFATWKNISGSIEIVKQGCWLDDINCYDRTDCIEKKDSPEVYFCCCEGNMCNEKFSYFPEM",  # noqa: E501
+            "resids": "38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101"  # noqa: E501
+            },
+        }
+    return results
 
 
-class Test_dssp_ss_saver:
-    """Test dssp secondary structure saver."""
-
-    dssp_fileD = tcommons.data_folder / '1ABC_D.dssp'
-    dssp_fileE = tcommons.data_folder / '1ABC_E.dssp'
-    
-    obj1 = libparse.DSSPParser(
-        fin=dssp_fileD,
-        pdbid='1ABC_D',
+def test_split_pdb_by_dssp(BTE_A_results):
+    """Test mkdssp with split."""
+    sgen = libparse.split_pdb_by_dssp(
+        Path(tcommons.data_folder, '1BTE_A.pdb'),
+        Path(tcommons.data_folder, '1BTE_A.dssp').read_text(),
+        reduced=True,
         )
-    obj2 = libparse.DSSPParser(
-        fin=dssp_fileE,
-        pdbid='1ABC_E',
+
+    _perform_dssp_split(sgen, BTE_A_results)
+
+
+def _perform_dssp_split(sgen, expected):
+    """
+    Separate logic to test PDB split by DSSP.
+
+    Logic is separated because there are two entry points:
+        * the direct function: split_pdb_by_dssp
+        * by mkdssp: mkdssp_w_split
+    """
+    counter = 0
+    for fout, d, pdbpart in sgen:
+        with open(tcommons.data_folder / f'{fout}.pdb', 'rb') as PART:
+            assert pdbpart == PART.read()
+
+        assert expected[fout] == d
+        counter += 1
+
+    assert counter == 2
+
+    with pytest.raises(StopIteration):
+        next(sgen)
+
+
+def test_split_pdb_by_dssp_minimum():
+    """Test mkdssp with split."""
+    results = {
+        # this seg0 actually corresponds to seg1 if `minimum` is not used
+        "1BTE_A_seg0": {
+            "dssp": "LLEEEEEEEEELLEEEEEEEEEELLLLLLLLLLLEEELLLLLLLEEEEELLLLHHHLEEELLLL",  # noqa: E501
+            "fasta": "RRHCFATWKNISGSIEIVKQGCWLDDINCYDRTDCIEKKDSPEVYFCCCEGNMCNEKFSYFPEM",  # noqa: E501
+            "resids": "38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101"  # noqa: E501
+            },
+        }
+
+    sgen = libparse.split_pdb_by_dssp(
+        Path(tcommons.data_folder, '1BTE_A.pdb'),
+        Path(tcommons.data_folder, '1BTE_A.dssp').read_text(),
+        reduced=True,
+        minimum=40,
         )
-    
-    def test_concat_dsspparsers_1(self):
-        """Test concatenation of DSSPParsers."""
-        output = libparse._concatenate_ss_from_dsspparsers([
-            self.obj1,
-            self.obj2,
-            ])
 
-        expected = ['1ABC_D| HGIBE TS ', '1ABC_E| HGIBE TS']
-        assert expected == output
+    counter = 0
+    for fout, d, pdbpart in sgen:
+        with open(tcommons.data_folder / '1BTE_A_seg1.pdb', 'rb') as PART:
+            assert pdbpart == PART.read()
+        counter += 1
+        assert results[fout] == d
 
-    def test_concat_dsspparsers_2(self):
-        """Test concatenation of DSSPParsers sorted output."""
-        output = libparse._concatenate_ss_from_dsspparsers([
-            self.obj2,
-            self.obj1,
-            ])
+    assert counter == 1  # make sure generator had only 1 element
 
-        expected = ['1ABC_D| HGIBE TS ', '1ABC_E| HGIBE TS']
-        assert expected == output
-    
-    def test_export_ss_1(self):
-        """Test export ss structure."""
-        libparse.export_ss_from_DSSP(
-            self.obj1,
-            self.obj2,
-            output=Path(tcommons.folder_output, 'dssp.database'),
+    with pytest.raises(StopIteration):
+        next(sgen)
+
+
+def test_mkdssp_w_split(BTE_A_results):
+    """."""
+    try:
+        sgen = libparse.mkdssp_w_split(
+            Path(tcommons.data_folder, '1BTE_A.pdb'),
+            'mkdssp',
+            reduced=True,
             )
-    
-        ofile = Path(tcommons.folder_output, 'dssp.database')
-        results = ofile.read_text()
+    except FileNotFoundError:
+        # in case mkdssp is not installed in the running machine
+        return
 
-        expected = '1ABC_D| HGIBE TS \n1ABC_E| HGIBE TS\n'
-        assert expected == results
-    
-    def test_export_ss_2(self):
-        """Test export ss to sys.stdout."""
-        libparse.export_ss_from_DSSP(
-            self.obj1,
-            self.obj2,
-            output=None,
-            )
+    _perform_dssp_split(sgen, BTE_A_results)
