@@ -1,180 +1,87 @@
 """Test libs for multicore operations."""
-import multiprocessing
-import subprocess
-
-import pytest
+import os
+from functools import partial
 
 from idpconfgen.libs import libmulticore as LM
 
-from . import tcommons
+
+def dummy_generator(a, b=None):
+    """Do dummy generator."""
+    for i in range(10):
+        yield i, a, b
 
 
-def test_Task_1():
-    """Test Task initiation."""
-    LM.Task()
+def test_consume_in_list():
+    """Test consume generator in list."""
+    results = LM.consume_iterable_in_list(dummy_generator, 'foo', b='bar')
+    for i, (j, a, b) in enumerate(results):
+        assert i == j
+        assert a == 'foo'
+        assert b == 'bar'
 
 
-class TestSubprocessTask:
-    """Test SubprocessTask."""
+def test_pool_function():
+    """Test pool function."""
+    execute = LM.pool_function(str, range(10_000), ncores=os.cpu_count())
+    result = list(execute)
+    assert all(isinstance(i, str) for i in result)
+    assert len(result) == 10_000
 
-    @pytest.mark.parametrize(
-        'in1,in2,expected',
-        [
-            ('ls', [tcommons.data_folder.str()], ['ls']),
-            (['ls'], [tcommons.data_folder.str()], ['ls']),
-            (['ls -ltr'], [tcommons.data_folder.str()], ['ls', '-ltr']),
-            ('ls -ltr', [tcommons.data_folder.str()], ['ls', '-ltr']),
-            (['ls', '-ltr'], [tcommons.data_folder.str()], ['ls', '-ltr']),
-            ],
+
+def test_pool_in_chunks():
+    """Test pool in chunks."""
+    execute = LM.pool_function_in_chunks(
+        str,
+        range(10_000),
+        ncores=os.cpu_count(),
+        chunks=200,
         )
-    def test_SubprocessTask_1(self, in1, in2, expected):
-        """Test cmd_exec from string."""
-        sub = LM.SubprocessTask(in1, in2)
-        assert sub.cmd_exec == expected
-
-    def test_SubprocessTask_2(self):
-        """Test prepare_cmd()."""
-        sub = LM.SubprocessTask('ls', [tcommons.data_folder.str()])
-        sub.prepare_cmd()
-        assert sub.cmd == ['ls', tcommons.data_folder.str()]
-
-    def test_SubprocessTask_3(self):
-        """Test execute."""
-        sub = LM.SubprocessTask('ls', [tcommons.data_folder.str()])
-        sub.prepare_cmd()
-        sub.execute()
-
-    def test_SubprocessTask_4(self):
-        """Test result is CompletedProcess."""
-        sub = LM.SubprocessTask('ls', [tcommons.data_folder.str()])
-        sub()
-        assert isinstance(sub.result, subprocess.CompletedProcess)
-
-    def test_SubprocessTask_5(self):
-        """Test string before prepare_cmd()."""
-        sub = LM.SubprocessTask('ls', [tcommons.data_folder.str()])
-        expected = "SubprocessTask(cmd_exec=['ls'],input=['{}'])".format(
-            tcommons.data_folder.str()
-            )
-        assert str(sub) == expected
-
-    def test_SubprocessTask_6(self):
-        """Test repr()."""
-        sub = LM.SubprocessTask('ls', [tcommons.data_folder.str()])
-        assert repr(sub) == \
-            "SubprocessTask(cmd_exec=['ls'],input=['{}'])".format(
-                tcommons.data_folder.str())
-
-    def test_SubprocessTask_7(self):
-        """Test str() after prepare_cmd()."""
-        sub = LM.SubprocessTask('ls', [tcommons.data_folder.str()])
-        sub.prepare_cmd()
-        assert str(sub) == 'ls {}'.format(tcommons.data_folder.str())
-
-    def test_SubprocessTask_8(self):
-        """Test input None."""
-        sub = LM.SubprocessTask('ls')
-        sub()
+    for chunk in execute:
+        assert len(chunk) == 200
+        assert isinstance(chunk, list)
+        assert isinstance(chunk[0], str)
 
 
-class TestDSSPTask:
-    """Test dedicated DSSP Task."""
-
-    @pytest.mark.parametrize(
-        'in1,in2',
-        [
-            ('dssp', '1XXX.pdb'),
-            (['dssp'], '1XXX.pdb'),
-            ],
+def test_pool_in_chunks_nested():
+    """Test pool in chunk from nested result."""
+    execute = LM.pool_function_in_chunks(
+        LM.consume_iterable_in_list,
+        range(20),
+        dummy_generator,
+        ncores=os.cpu_count(),
+        chunks=5,
         )
-    def test_DSSPTask_1(self, in1, in2):
-        """Test DSSPTask init."""
-        LM.DSSPTask(in1, in2)
 
-    def test_DSSPTask_2(self):
-        """Test raises ValueError when input is missing."""
-        with pytest.raises(TypeError):
-            LM.DSSPTask('dssp')
-  
-    @pytest.mark.parametrize(
-        'in1,in2,expected',
-        [
-            ('dssp', '1XXX.pdb', ['dssp', '-i', '1XXX.pdb']),
-            (['dssp'], '1XXX.pdb', ['dssp', '-i', '1XXX.pdb']),
-            ]
+    for chunk in execute:
+
+        # chunk of results
+        assert len(chunk) == 5
+        assert isinstance(chunk, list)
+
+        # this is a list containing the results from `dummy_generator`
+        # list of tuples, contains 10 elements because `dummy_generator`
+        # yields 10x
+        assert isinstance(chunk[0], list)
+        assert len(chunk[0]) == 10
+
+        # each yielded results is a tuple of 3 elements
+        assert isinstance(chunk[0][0], tuple)
+        assert len(chunk[0][0]) == 3
+
+
+def test_pool_in_chunks_flatten():
+    """Test pool in chunk from nested result."""
+    execute = partial(
+        LM.pool_function_in_chunks,
+        LM.consume_iterable_in_list,
+        range(20),
+        dummy_generator,
+        ncores=os.cpu_count(),
+        chunks=5,
         )
-    def test_DSSPTask_3(self, in1, in2, expected):
-        """Test DSSPTask prepare_cmd."""
-        dssptask = LM.DSSPTask(in1, in2)
-        dssptask.prepare_cmd()
-        assert dssptask.cmd == expected
 
-    def test_DSSPTask_pdb_path(self):
-        """Test DSSPTask pdb_path attribute."""
-        dssptask = LM.DSSPTask('dssp', '1XXX.pdb')
-        assert dssptask.pdb_path == '1XXX.pdb'
+    def assrt(flat):
+        for result in flat:
+            assert isinstance(result, tuple)
 
-    def test_DSSPTask_call(self):
-        """Test DSSPTask call dunder."""
-        dssptask = LM.DSSPTask('ls', '-ltr')
-        dssptask()
-
-
-class TestWorker:
-    """Test Worker."""
-
-    def test_worker_1(self):
-        """Test Worker init."""
-        LM.Worker(None, None)
-
-    def test_worker_2(self):
-        """Test Worker raises TypeError."""
-        with pytest.raises(TypeError):
-            LM.Worker()
-
-    def test_worker_3(self):
-        """Test Worker run."""
-        tasks = multiprocessing.JoinableQueue()
-        queue = multiprocessing.Queue()
-        workers = [LM.Worker(tasks, queue)]
-        for w in workers:
-            w.start()
-        tasks.put(LM.SubprocessTask('ls -ltr'))
-        tasks.put(None)
-        tasks.join()
-        numjobs = 1
-        while numjobs:
-            queue.get()
-            numjobs -= 1
-
-
-class TestJoinedResults:
-    """Test JoinedResults multicore operator."""
-
-    def test_init(self):
-        """Test JoinedResults inits with correct attributes."""
-        jr = LM.JoinedResults(
-            '-ltr',
-            'ls',
-            ncores=7,
-            TaskMethod=LM.Task,
-            results_parser=int)
-        assert jr.input_data == '-ltr'
-        assert jr.cmd == 'ls'
-        assert jr.ncores == 7
-        assert jr.TaskMethod == LM.Task
-        assert jr.rp == int
-
-    def test_build(self):
-        """Test JoinedResults build() method."""
-        jr = LM.JoinedResults('-ltr', 'ls', ncores=7)
-        jr.build()
-        assert isinstance(jr.tasks, multiprocessing.queues.JoinableQueue)
-        assert isinstance(jr.queue_results, multiprocessing.queues.Queue)
-        assert len(jr.workers) == 7
-        assert all(isinstance(w, LM.Worker) for w in jr.workers)
-
-    def test_run(self):
-        """Test JoinedResults run() method."""
-        jr = LM.JoinedResults('-ltr', 'ls', ncores=7)
-        jr.run()
+    LM.flat_results_from_chunk(execute, assrt)
